Beginning fixes to lighting codepath
We need a way to draw to a GR1616 image format texture to make ShadowMapResolver work properly. It depends on having 2 channels of 16-bit floating point precision instead of 4 channels of 8-bit. We have an enum which supports this, but SFML itself doesn't support it. We'll have to use OpenGL directly to accomplish it.