Adds basic Lua scripting.
Powered by NLua.

Made a new component, ScriptComponent for this purpose. It has 2
parameters:

* Code: Lua code to execute on creation.
* File: file in SS14.Server/Scripts/ to run as Lua.

If the a file is loaded more than once it will be loaded from an internal
cache to prevent overhead from file I/O.

If the component receives any messages and has a function with the same
name in the lua instance, that function will be called.