Fixes exception with map saving
This exception would occur due to the map saving system (F5) taking the current name of the server and saving over the current server's map. This would cause an accessdenied exception due to the map being in use while the server is running presumably. This fixes that problem and adds a timestamp added when saving maps using the debug verb (to prevent any overwriting of maps.