Allow components to be created while iterating them.
This is necessary to avoid tons of papercuts with entity systems.
Without this it is not possible to create entities inside the update loop of an entity system, which is too common for me to disallow.

As for performance, I offset it by removing the LINQ usage.
It's marginally faster than before even.