Improve object printing in VV
Adds the `PrettyPrint` class, which should help with producing object
representations for VV and other debug menus.

The new class prints generics more nicely, abbreviating type arguments
and not printing them with the assembly name in.

A common pattern was to print an object and its type, but only if the
object didn't look like its type. That now has a method to do it for
you.