Add system for using flag names in YAML
It's now possible to hook an enum into the serialization and
deserialization of a field by name, using an attribute. This allows for
bitflags in the engine to be defined by name in the content, in a way
that lets content creators write flag names in prototype YAML.

This works by annotating an enum with the [FlagsFor(<fieldname>)]
attribute, which the YamlObjectSerializer picks up through the
ReflectionManager to use as a special serializer for fields of that
name.