Fix Robust.Analyzer foobar test case, improve messages
Saw a trick in Robust.UnitTesting to load the test assembly for
reflection, which allows Robust.Analyzer.Test to define its own
prototype type for testing purposes. This fixes the foobar test case and
removes the hardcoded ignore in Robust.Analyzer.

Also improves some messages to clarify the prototype type being checked
for, including a message which contained "entity" incorrectly.