Fix lasers changing DrawingHandleWorld transform
Lasers were manipulating the global drawing transform
instead of just rendering in the correct direction. They
have been changed to use a rotated box.

Fixed an issue with Clyde not respecting the way Box2Rotated
rotates, which was not allowing for the behaviour needed to
draw lasers. This has added some overhead to rendering, but
hopefully not too much. It can always be optimized later.