Fix GameTiming tests, starting timing from first tick
Timing was being started from tick 0, but this was causing problems
because the default tickrate is 60, so the "first" tick was always
1/60th of a second long. Now, it starts from tick 1, so the "first" tick
takes no time.

Fixes some "off-by-one" errors in the tests caused by this change. They
were expecting that, at tick 60, 60 ticks would have passed. Previously,
that wasn't the case, but the timing system made it appear so. Now, they
correctly use the 61st tick instead.