Cleanup contact code a lot
A lot of this stuff is farseer porting box2d verbatim.
Most of the cleanup is just doing it using existing dotnet stuff.
I also moved contact storage directly onto fixtures which box2d has an open issue for.
This removes the particularly slow O(n) check for AddPair but that really only matters with lots of contacts on a body.
Removed a bunch of refs from stuff as a result.