Fix window blit synchronization issues.
We were trying to load the sync object off the wrong render target (the internal RenderTexture from the GL context manager, not the RenderWindow that's exposed). This meant sync wasn't happening.

OpenGL was reporting errors because the sync object was 0, but because debug output wasn't enabled in the secondary OpenGL contexts this was falling on deaf ears.

Woops.