Don't run client physics in frameupdate
Help wanted.

So AFAIK the only issue is that lerping only works for server states coming in, not client physics. Essentially what we need to do is whenever client sets LocalPosition of an entity it attempts to lerp it from A to B. This should be separate I think from server lerping where we may not want to linearly lerp.