Add entitylookup for components
Technically entitylookup is now just "lookup" but changing it would be a pain.
I used the same API as entitylookup and it's kinda boilerplatey but don't know of an easier way to do it. The main advantage is it internally checks if it shoud just iterate every component directly or check the dynamic trees. This also means content has a consistent way to do component checks even if we change the internal implementation as right now people call the entitylookup methods and use their own entityquery.