Fix broadphase desync
So this is another fun MAPLOADER ONE.

The issue is because GridInitialize is being used for the broadphase there's ordering bugs where the grid is being initialized AFTER all these bodies are being added to the (map's) broadphase. The workaround for now is to just do it in Add instead (until maploader properly goes top-down).