Bandaid maploader for loading onto an existing one
I've outlined my thoughts on what maploader should be elsewhere but being able to just change the root node to null would be very useful instead of having some entities partially initing on an entity that will be deleted.

This could probably be a bit cleaner but it's still not going to be the ideal solution IMO so as long as it works for now.