Make RobustServerSim public
Okay so a long time ago in a maintainer meeting I went on a rant.

The issue is we have the following ways to run tests:
- Actual unit test
- RobustUnitTest
- RobustServerSim
- RobustIntegrationTest
- ContentIntegrationTest

Ideally we would kill RobustUnitTest because RobustServerSim is the more recent, less poorly coded, and faster version that can also be re-used for benchmarks.