Adjust centerpoint on non 32px sprits if offset = false
This is to fix an issue with setting SpriteOffset to false on sprites that are larger than 32px.

SpriteOffset = false zero'ed the vec2 in the draw enitity, however this was never tested as SpriteOffset=false has not been used by any content. It doesn't work as expected and this is designed to fix that.

For 32px this will set to 0,1. if 64 will be 0,0.5, if 96 will be 0,0.33 etc...