Apply entity states in order
Doesn't seem to break anything?

The issue is due to some internal transform ordering that happened at least a month+ ago afaict. Because the entity state for the parent (the grid) isn't applied yet then the child doesn't have the correct griduid to use.

By applying the parent first (instead of in a separate step) it ensures all the data is available.