Allow deletion of FileLogHandler logs while engine is running
This is prompted by a warning in TGS with OpenDream.

TGS has an option to either preserve or delete engine logs. It tries to delete the log immediately after the process terminates, but before the `System.Diagnostics.Process` is disposed. On Windows, this fails due to the handle still being considered open.

Create the FileStream manually instead. Parameters are exactly the same as before with the exception that the handle now allows other processes to delete the file while it's open.