Rewrite memory management approach.
The previous iteration used a single massive POH array to store the data for all players (and also the metadata cache). References to this array were in the form of direct unmanaged pointers into this array.

This was both incredibly dangerous (raw pointers in complex multithreaded code) and incredibly complex do to management of growth and resizes.

This rewrites all that to do the following:
1. Use a separate memory allocation per player
2. Pass around bounds-checked int offsets instead of raw pointers.
3. Use a native allocation instead of raw pointers.
4. (Use advanced virtual memory shenanigans to handle grow/shrink of the memory regions)

Going through those in a weird order:

For (3), I didn't want to trust the GC to (a) avoid populating memory unnecessarily when allocating the objects and (b) have to deal with significant GC pressure in the POH on player reconnect.

The easiest way to explain why we're doing (4) is because I got really excited to do something *cool*. Yeah quite frankly this was more justified in my head initially but looking at it now I could've just use NativeMemory.Realloc() and called it a day. Ah well.

(2) just massively reduces the complexity of the memory management. Now we don't need a whole allocation system (and state machinery) for player allocations anymore.

(1) makes the entire code *far* more safe against bugs and, worst case, makes bugs corrupt PVS state instead of corrupting the entire process. This does theoretically have minor performance overhead, but the cost of bounds checks is probably insignificant against the main savings of the PR (removing dictionary use).

Like seriously while I was untangling this code I was just outright scared about some of the complex intertwined ways the pointers and memory management and all that stuff played together. I'm very glad to get rid of it.

Other relatively minor changes below:

The size of the Pvs(Meta)Data struct has been padded to 16 bytes. This makes sure they don't split cache lines and some memory operations are more efficient.

The pool of free indexes has been turned into a linked list stored in the unused metadata slots themselves. This avoids needing a separate stack.

All PVS indexes are now a PvsIndex type to distinguish them.

The virtual memory region management is its own ResizableMemoryRegion class, (ab)use to your heart's content.

Added a MathHelper.CeilMultipleOfPowerOfTwo helper because I needed it for ResizableMemoryRegion.